requirements.txt
black==22.10.0
cloudpickle==1.3.0
dm-control==0.0.403778684
glfw==2.5.4
gym==0.23.0
imageio==2.21.2
mani-skill2-0.4.2
matplotlib==3.4.2
metaworld==0.0.1.dev0
moviepy==1.0.3
mujoco-py==2.1.2.14
numpy==1.23.2
patchelf==0.17.2.1
pybullet==3.2.5
scikit-video==1.1.11
tensorboard==2.6.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.0
wandb==0.15.3